Coffee Machine
This is meant to act as a coffee machine. The user can choose a latte, cappuccino, or espresso. the machine has certain resources such as milk, water, and coffee. Resources are deducted when the user buys a coffee. The user pays with coins and the machine calculates whether the user has paid enough and what change the user is owed.